Refactor text-to-speech module to use Piper TTS instead of pyttsx3
- Removed pyttsx3 dependency and related threading logic for speech synthesis.
- Integrated Piper TTS for voice loading and audio playback using sounddevice.
- Updated voice listing to reflect available .onnx files in the model path.
- Simplified voice settings retrieval and synthesis configuration.
- Adjusted speak_text function to handle audio playback and interruption with Piper TTS.

+++ desktop/app.py
@@ -253,7 +253,7 @@ class Api:
         _speak_in_background(
             "This is how I sound.",
             voice_id=voice_id or None,
-            rate=int(rate) if rate else None,
+            rate=float(rate) if rate else None,
             volume=float(volume) if volume not in (None, "") else None,
         )
         return {"ok": True}
